Update logic in loadRuntimeEnv to check if code is executing in the browser
If this module is used in server side rendering it attempts to parse runtimeEnv
and it will fail due to the value of runtimeEnv still containing the place holder.
The parsing of the runtimeEnv should only happen in the browser as that is where
the place holder will be replaced.

This should then work correctly for both server side rendering and browser rendering
in production as both options will use the production config vars as opposed to the
compiled configuration values.